297: Assign company to a customer registered after placing the order

+++ Service/Customer.php
@@ -82,7 +82,7 @@ class Customer
      *
      * @return int|null
      */
-    public function getCustomerAmountAvailable(): ?int
+    public function getCustomerAmountAvailable(): ?float
     {
         $amountAvailable = null;
         if ($customerId = $this->customerSession->getCustomerId()) {
